Split proxies' py files to its own module
